Muevo funciones de distribution fuera de helpers
