+++ ckanext/gobar_theme/helpers/__init__.py
@@ -4,3 +4,5 @@ from .organizations import *
 from .groups import *
 from .licenses import *
 from .types import *
+from .cron import *
+from .distribution import *
